+++ bin/rdpy-vncscreenshot.py
@@ -24,10 +24,10 @@ take screenshot of login page
 """
 
 import sys, os, getopt
-from PyQt4 import QtCore, QtGui
+from PyQt5 import QtCore, QtGui, QtWidgets
 from rdpy.protocol.rfb import rfb
 import rdpy.core.log as log
-from rdpy.ui.qt4 import qtImageFormatFromRFBPixelFormat
+from rdpy.ui.qt5 import qtImageFormatFromRFBPixelFormat
 from twisted.internet import task
 
 #set log level
@@ -155,11 +155,11 @@ if __name__ == '__main__':
             password = arg
         
     #create application
-    app = QtGui.QApplication(sys.argv)
+    app = QtWidgets.QApplication(sys.argv)
     
-    #add qt4 reactor
-    import qt4reactor
-    qt4reactor.install()
+    #add qt5 reactor
+    import qt5reactor
+    qt5reactor.install()
     from twisted.internet import reactor
